feat: implement driver Start command
Also refactored/re-ordered the functions in driver.go to be better laid
out with similar functions together, and moved non driver functions to
util.go.
